requirements.txt
fastapi
uvicorn
boto3
python-dotenv
requests
streamlit
python-multipart
pymilvus